wip: moved app to appshell/ (still fixing rollup exports)
