+++ static_report/src/appshell/index.tsx
@@ -2,8 +2,8 @@ import { ChakraProvider } from '@chakra-ui/react';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App';
-import reportWebVitals from './utils/reportWebVitals';
-import theme from './utils/theme';
+import reportWebVitals from '../utils/reportWebVitals';
+import theme from '../utils/theme';
 import 'inter-ui/inter-latin.css';
 
 const rootElement = document.getElementById('root') as HTMLElement;
