+++ static_report/src/components/shared/Layouts/index.ts
@@ -1,9 +1,9 @@
 export * from './BreadcrumbNav';
 export * from './CollapseContent';
-export * from './Footer';
+// export * from './Footer';
 export * from './Loading';
-export * from './Main';
-export * from './Navbar';
+// export * from './Main';
+// export * from './Navbar';
 export * from './NoData';
 export * from './NotFound';
-export * from './Sidebar';
+// export * from './Sidebar';
