+++ static_report/src/index.ts
@@ -1,3 +1,6 @@
+/**
+ * Main Entry File for UI Library
+ */
 export * from './components/shared';
 export * from './utils';
 export * from './hooks';
